requirements.txt
Django==4.2.7
Pillow==10.1.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0 
psycopg2-binary==2.9.9
dj-database-url==2.1.0
xlsxwriter==3.2.0
reportlab==4.2.0 